linkdeal_app/backend/LinkDeal/requirements.txt
Django==5.2.8
djangorestframework==3.16.1
python-dotenv==1.2.1
django-cors-headers==4.9.0
psycopg2-binary==2.9.11
python-jose[cryptography]==3.5.0
Pillow==12.0.0
requests==2.32.5
gunicorn==23.0.0
whitenoise==6.8.2
APScheduler==3.10.4
numpy==1.26.4
openai==1.60.1

